refactor: migrated dataset/preview/style.json
