+++ sites/geohub/src/api/datasets/[id]/permission/POST.ts
@@ -42,7 +42,7 @@ export const Error = {
 	403: appError(403, 'Permission error')
 };
 
-export default new Endpoint({ Param, Output, Modifier }).handle(
+export default new Endpoint({ Param, Output, Modifier, Error }).handle(
 	async (param, { locals, request }) => {
 		const session = await locals.auth();
 		if (!session) error(403, { message: 'Permission error' });
